refactor(firmware): split net/http.rs response helpers into respond submodule
http.rs was 2278 LoC. Extracts the leaf response helpers into a sibling
respond submodule so the route handlers stay focused on request dispatch.

Moves to crates/stackchan-firmware/src/net/respond.rs:
- HttpError enum (pub(super))
- DASHBOARD_GZ const + write_dashboard
- write_no_content, write_json, write_text, write_status_for_error,
  write_unauthorized, status_reason
- health_body, state_body, sensors_body, tasks_body, events_body,
  json_escape

http.rs drops 2278 to 1956 LoC; respond.rs lands at ~360 LoC.
Visibility is pub(super) so no external API change.

Validated via cargo +esp clippy --release -- -D warnings and
cargo +esp build --release.
